+++ src/builtins/builtins.c
@@ -1,4 +1,4 @@
-#include "modules.h"
+#include "builtins.h"
 
 #include "core.h"
 #include "core.jsc.inc"
